+++ src/roq/binance_futures/drop_copy.cpp
@@ -324,7 +324,7 @@ void DropCopy::operator()(Trace<json::OrderTradeUpdate> const &event) {
             .update_type = {},
             .user = {},
         };
-        create_trace_and_dispatch(handler_, trace_info, trade_update, true, SOURCE_NONE);
+        create_trace_and_dispatch(handler_, trace_info, trade_update, true, SOURCE_SELF);
       } else {
         log::warn("*** EXTERNAL ORDER ***"sv);
         log::warn("execution_report={}"sv, execution_report);